requirements.txt
bitarray==0.8.3
Pillow==5.4.1
